moving test_to_pi_string to new test_viz file
