+++ tests/unit/test_gkp.py
@@ -15,37 +15,15 @@
 
 # pylint: disable=no-self-use
 
-import math
-
 import numpy as np
 from numpy import sqrt, pi
 from numpy.random import default_rng as rng
 import pytest
 
-from flamingpy.cv.gkp import to_pi_string, integer_fractional, GKP_binner, Z_err, Z_err_cond
+from flamingpy.cv.gkp import integer_fractional, GKP_binner, Z_err, Z_err_cond
 
 
 N = 50
-
-
-def test_to_pi_string():
-    """Test for the convenience function to_pi_string."""
-    # Test +- sqrt(pi) and sqrt(pi)/2.
-    assert to_pi_string(np.sqrt(np.pi)) == "$\\sqrt{\\pi}$"
-    assert to_pi_string(-np.sqrt(np.pi)) == "$-\\sqrt{\\pi}$"
-    assert to_pi_string(np.sqrt(np.pi) / 2) == "$\\sqrt{\\pi}/2$"
-    assert to_pi_string(-np.sqrt(np.pi) / 2) == "$-\\sqrt{\\pi}/2$"
-    # Test random odd integer multiples of sqrt(pi)/2, eccept 1 and -1
-    odd_int = (2 * rng().integers(2, 25) - 1) * (-1) ** rng().integers(2)
-    assert to_pi_string(odd_int * np.sqrt(np.pi) / 2) == "${}\\sqrt{{\\pi}}/2$".format(odd_int)
-    #  Test random even multiples of sqrt(pi).
-    even_int = odd_int + 1
-    assert to_pi_string(even_int * np.sqrt(np.pi)) == "${}\\sqrt{{\\pi}}$".format(even_int)
-    # Check everything else converted into a str.
-    rand_numb = rng().random()
-    if not np.isclose(math.remainder(rand_numb, np.sqrt(np.pi) / 2), 0):
-        assert to_pi_string(rand_numb) == str(rand_numb)
-
 
 # Construct random numbers from an integer and fractional part.
 alpha_vals = np.append(np.random.rand(5) * 5, np.sqrt(np.pi))
